refactor: 퀴즈 관련 응답 변환 Mapper 분리
